add MaxDepth in CONF
Reducing time/inc causes problems eventually:
- communication/processing overhead
- noise from the OS and other processes, task switching etc.
- timer resolution, OS dependant (typical 10ms)
Using a depth limit avoids these problems, and allows to reliably play
very fast games.

On the other hand, a depth limit can be dangerous too:
- a single search explosion can block for avery long time
- the depth may be too large or too small depending on the position
So combining the two is useful.

Typical example, when tuning eval parameters that are not really depth
dependant, would be something like tc=2+0.02 depth=10 or so. In my
experience this kind of combination gives good results, in the sense
that it gives high drawelo for a given testing throughput.